backend/apps/accounts/views.py
from rest_framework import generics, permissions
from rest_framework.response import Response

from django.contrib.auth import get_user_model
from .serializers import RegisterSerializer, MeSerializer, MeUpdateSerializer

User = get_user_model()

class RegisterView(generics.CreateAPIView):
    permission_classes = [permissions.AllowAny]
    serializer_class = RegisterSerializer

class MeView(generics.GenericAPIView):
    permission_classes = [permissions.IsAuthenticated]

    def get(self, request):
        return Response(MeSerializer(request.user).data)

    def patch(self, request):
        serializer = MeUpdateSerializer(data=request.data, context={"request": request})
        serializer.is_valid(raise_exception=True)
        serializer.update(request.user, serializer.validated_data)
        return Response(MeSerializer(request.user).data)
